fixed: A few small :gcc warnings
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/linuxport/XBMC@11402 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/lib/libGoAhead/XBMCweb.cpp
@@ -375,7 +375,7 @@ int CXbmcWeb::xbmcNavigatorState( int eid, webs_t wp, char_t *parameter)
     while( xbmcNavigator[cmd].xbmcAppStateCode != -1)
     {
       // if we have a match
-      if(xbmcNavigator[cmd].xbmcAppStateCode == navigatorState)
+      if((DWORD)xbmcNavigator[cmd].xbmcAppStateCode == navigatorState)
       {
         if( eid != NO_EID) {
           ejSetResult( eid, xbmcNavigator[cmd].xbmcNavigateParameter);
